pyproject.toml
[project]
name = "llm-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "google-genai>=1.13.0",
    "google-generativeai>=0.8.5",
    "gunicorn>=23.0.0",
    "langchain-google-genai>=2.0.10",
    "langgraph>=0.6.2",
    "openai>=1.77.0",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "redis>=6.0.0",
    "sqlalchemy>=2.0.40",
    "tzdata>=2025.2",
    "uvicorn>=0.34.2",
]
